requirements.txt
psycopg2-binary
flask
gunicorn
python-dotenv